Bump base-compat and hashable upper bounds
There's no local testing target for this, just a complicated Travis
setup, so I'll rely on merge CI checks for correctness. I forced use
of this on one of my projects with Stackage LTS 24, and it seemed to
at least compile, so let's give it a try. Would be good to document
testing procedure for upgrades.